+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/workrecord/erroralarm/condition/attendanceitem/KrcstErAlSingleFixed.java
@@ -4,7 +4,6 @@
 package nts.uk.ctx.at.record.infra.entity.workrecord.erroralarm.condition.attendanceitem;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 import javax.persistence.Basic;
 import javax.persistence.Column;
@@ -37,7 +36,7 @@ public class KrcstErAlSingleFixed extends UkJpaEntity implements Serializable {
 	@Basic(optional = false)
 	@NotNull
 	@Column(name = "FIXED_VALUE")
-	public BigDecimal fixedValue;
+	public int fixedValue;
 	
 	@OneToOne
 	@JoinColumns({
@@ -49,7 +48,7 @@ public class KrcstErAlSingleFixed extends UkJpaEntity implements Serializable {
 	protected Object getKey() {
 		return this.krcstEralSingleFixedPK;
 	}
-	public KrcstErAlSingleFixed(KrcstErAlSingleFixedPK krcstEralSingleFixedPK, BigDecimal fixedValue) {
+	public KrcstErAlSingleFixed(KrcstErAlSingleFixedPK krcstEralSingleFixedPK, int fixedValue) {
 		super();
 		this.krcstEralSingleFixedPK = krcstEralSingleFixedPK;
 		this.fixedValue = fixedValue;
